Data Retrieval, Visualization, and Styling
In this assignment, I retrieved the data from the API link and used JSON to convert that retrieved data into Objects that is readable and understandable. Data retrieved and parsed by JSON is visualized in form of table with its heading. Styling were applied to this webpage to differentiate color of each element in the page.